Refactoring de l'architecture du projet
