+++ Java_Rectangle/src/forms/Union.java
@@ -1,4 +1,4 @@
-package simpleform;
+package forms;
 
 import java.awt.Graphics;
