+++ Java_Rectangle/src/forms/icons/diff.png
