+++ Java_Rectangle/src/forms/icons/union.png
