+++ Java_Rectangle/src/oldform/Difference.java
@@ -1,4 +1,4 @@
-package form;
+package oldform;
 
 import java.awt.Color;
 import java.awt.Graphics;
