+++ Java_Rectangle/src/oldform/Intersection.java
@@ -1,4 +1,4 @@
-package form;
+package oldform;
 
 import java.awt.Graphics;
