+++ Java_Rectangle/src/oldform/Shape.java
@@ -1,4 +1,4 @@
-package form;
+package oldform;
 
 import java.awt.Graphics;
 import java.io.Serializable;
