+++ Java_Rectangle/src/oldform/TerminalDemo.java
@@ -1,4 +1,4 @@
-package form;
+package oldform;
 
 import java.beans.XMLDecoder;
 import java.beans.XMLEncoder;
